Almost Finished
Everything is functioning with exception to Mutation Rate and Cross
Rate. These two values do not change yet. They need to go up for any
value that does not perform change action and down to normal level after
every change action. With Cross Rate it needs to total out for the
Chromosome as well.